move fetcher into unixfs
