+++ src/unixfs/mod.rs
@@ -1,7 +1,7 @@
-use crate::block::Cid;
+use crate::block::{Block, Cid};
 use crate::error::Error;
 use crate::ipld::{Ipld, IpldDag, IpldPath, formats::pb::PbNode};
-use crate::repo::RepoTypes;
+use crate::repo::{BlockLoader, RepoTypes};
 use core::future::Future;
 use futures::compat::*;
 use std::collections::HashMap;
@@ -9,6 +9,7 @@ use std::convert::TryInto;
 use std::path::PathBuf;
 
 pub mod unixfs;
+use unixfs::{Data_DataType, Data as UnixfsData};
 
 pub struct File {
     data: Vec<u8>,
@@ -69,6 +70,76 @@ impl Into<String> for File {
     fn into(self) -> String {
         String::from_utf8_lossy(&self.data).to_string()
     }
+}
+
+
+pub(crate) fn extract_block(block: Block) -> Vec<u8> {
+    match Ipld::from(&block) {
+        Ok(ref i) => extract_file_data(i),
+        _ => None
+    }.unwrap_or_else(move || {
+        let (_cid, data) = block.into_inner();
+        data
+    }
+}
+
+pub(crate) fn extract_file_data<'d>(ipld: &'d Ipld) -> Option<Vec<u8>> {
+    match ipld {
+        Ipld::Bytes(b) => {
+            protobuf::parse_from_bytes::<UnixfsData>(&b)
+                .map(|mut u| u.take_Data()).ok()
+        },
+        Ipld::Object(hp) => match hp.get(&"Data".to_owned()) {
+                Some(i) => extract_file_data(i),
+                _ => None
+            },
+        _ => None
+    }
+}
+
+pub(crate) async fn fetch_file<'d, B: 'd + BlockLoader>(ipld: &'d Ipld, loader: B)
+    -> Result<Vec<u8>, Error>
+{
+    if let Ipld::Object(hp) = ipld {
+        if let Some(Ipld::Bytes(d)) = hp.get(&"Data".to_owned()) {
+            if let Ok(item) = protobuf::parse_from_bytes::<UnixfsData>(d) {
+                if item.get_Type() == Data_DataType::File && item.get_blocksizes().len() > 0
+                {
+                    if let Some(Ipld::Array(links)) = hp.get(&"Links".to_owned()) {
+                        let name = &"".to_owned();
+                        let mut buffer : Vec<u8> = vec![];
+                        // TODO: can we somehow return this iterator to hyper
+                        //       and have it chunk-transfer the data?
+                        for cid in links
+                            .iter()
+                            .filter_map(move |x|
+                                if let Ipld::Object(l) = x {
+                                    match (l.get(&"Name".to_owned()), l.get(&"Hash".to_owned())) {
+                                        (Some(Ipld::String(n)),
+                                            Some(Ipld::Cid(cid))) if n == name => Some(cid),
+                                        _=> None
+                                    }
+                                } else {
+                                    None
+                                }
+                        ) {
+                            let block = await!(loader.load_block(cid))?;
+                            let mut data = extract_block(block);
+                            buffer.append(&mut data)
+                        }
+
+                        return Ok(buffer)
+                    }
+                }
+
+                return Ok(item.get_Data().into())
+            } else {
+                return Ok(d.to_vec())
+            }
+        }
+    };
+
+    Err(Error::from(::std::io::Error::new(::std::io::ErrorKind::NotFound, "")))
 }
 
 #[cfg(test)]
